IPLD Refactor
Refactored the entire codebase to focus on IPLD based Merkle Tree
construction. This differs in several ways from the original Hash
based Merkle Tree construction, as it uses IPLD blocks and a
blockstore to "hash" and "merge" respectively.

This has removed a significant chunk of functionality from the
original implementation. Most notably, all the parallel code.

Additionally, the proof construction now relies soley on the "path"
parameter to traverse through the IPLD merkle tree from the
provided root to the target leaf. This creates *much* smaller
proofs as we don't have to store the siblings and their CIDs
(32 bytes) * height, and instead store a single uint32 path.

TODO:
- Re-add benchmarking
- More tests
- Re-enable build modes (only proof-gen supported)